Work in progress: MacroStategy Environment
In this env, the agent selects one of a few custom spells every time
step. The spells are scripted to send waves of different unit types
automatically toward the enemy.

A good agent will learn to observe the enemy's unit composition and send
waves of the corresponding unit type to win the battle.