Add immortals, defending probes
The reward structure is as follows:
        No-op yields 0 reward.
        Casting psi-storm or shield generates -50 reward (mana cost)
        Damaging a zergling generates +1 reward
        Killing a zergling generates +25 reward
        Losing a probe costs -100 reward

The AI should first learn that casting spells is bad, and no-op.
Eventually it should learn that casting spells is good, because
it damages the enemy zerglings.
Then it should learn to cast the combo of force field and psi-storm.
Finally, the AI should learn *not* to cast field/storm when an immortal
is present, because it wastes mana.